Hook up all landing components and add transition animations.
