+++ src/components/landing/LoginCard.jsx
@@ -11,12 +11,13 @@ function LoginCard({ switchCard }) {
   }
 
   return (
-    <LandingCard title="Welcome Back">
-      <form className="card-form" onSubmit={onSubmit}>
+    <LandingCard title="Welcome Back, Please Log In">
+      <form onSubmit={onSubmit}>
         <TextInput placeHolder="Username" name="Username" />
         <TextInput placeHolder="Password" name="Password" type="password" />
         <Button type="submit">Login</Button>
-        <Button onClick={() => switchCard('groupLogin')}>Have a group code?</Button>
+        <Button className="btn-link" onClick={() => switchCard('groupLogin')}>Have a group code?</Button>
+        <Button className="btn-link" onClick={() => switchCard('register')}>Don&apos;t have an Account?</Button>
       </form>
     </LandingCard>
   );
