+++ src/scss/_variables.scss
@@ -14,7 +14,7 @@ $variant-colors: (
 $box-shadows: (
   'sm': rgba(0, 0, 0, 0.35) 0px 5px 15px,
   'md': (rgba(50, 50, 93, 0.25) 0px 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px),
-  'lg': rgba(0, 0, 0.9, 0.1) 2px 5px 60px,
+  'lg': (rgba(0, 0, 0, 0.2) 0px 30px 90px),
 );
 
 $profile-length: 70px;
