+++ src/scss/animation.scss
@@ -1,0 +1,20 @@
+.fade-enter {
+  opacity: 0;
+  transform: translateY(-30%) scaleX(0.6) scaleY(0.6);
+}
+.fade-enter-active {
+  opacity: 1;
+  transform: translateY(0%) scaleX(1) scaleX(1);
+}
+.fade-exit {
+  opacity: 1;
+  transform: translateY(0%) scaleX(1) scaleY(1);
+}
+.fade-exit-active {
+  opacity: 0;
+  transform: translateY(30%) scaleX(0.6) scaleY(0.6);
+}
+.fade-enter-active,
+.fade-exit-active {
+  transition: opacity 350ms, transform 350ms;
+}
