+++ src/scss/register-card.scss
@@ -21,11 +21,11 @@
   }
 
   .btn-submit,
-  .btn-login {
+  .btn-link {
     width: 100%;
   }
 
-  .btn-login {
+  .btn-link {
     margin-top: 12px;
     background-color: transparent;
